+++ statickg/services/fuseki.py
@@ -290,12 +290,26 @@ class FusekiDataLoaderService(BaseFileService[FusekiDataLoaderServiceInvokeArgs]
 
         name = f"fuseki-{dbinfo.dir.name}"
         port = find_available_port(self.hostname, 3031)
-        (subprocess.check_output if self.capture_output else subprocess.check_call)(
-            self.get_start_command(args).format(
-                ID=name, PORT=str(port), DB_DIR=dbinfo.dir
-            ),
-            shell=True,
-        )
+        try:
+            (subprocess.check_output if self.capture_output else subprocess.check_call)(
+                self.get_start_command(args).format(
+                    ID=name, PORT=str(port), DB_DIR=dbinfo.dir
+                ),
+                shell=True,
+            )
+        except:
+            # sometime the process terminates but it leaves some trash behind -- stop and start again
+            (subprocess.check_output if self.capture_output else subprocess.check_call)(
+                self.get_stop_command(args).format(ID=name),
+                shell=True,
+            )
+            (subprocess.check_output if self.capture_output else subprocess.check_call)(
+                self.get_start_command(args).format(
+                    ID=name, PORT=str(port), DB_DIR=dbinfo.dir
+                ),
+                shell=True,
+            )
+
         self.started_services[dbinfo.dir] = (name, port)
         assert dbinfo.hostname is None
         dbinfo.hostname = f"{self.hostname}:{port}"